finance-chatbot/pyproject.toml
[project]
name = "finance-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=6.1.0",
    "faiss-cpu>=1.11.0",
    "flask>=3.1.1",
    "huggingface-hub>=0.33.4",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.27",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
]
